refactor(notifications): fix eslint warnings
