+++ services/notifications/src/components/accounts/templates/ChangePasswordTemplate.ts
@@ -1,4 +1,4 @@
-import { CLIENT_HOST } from '../../../configs/AppConfig'
+import { CLIENT_HOST } from '../../../configs/app-config'
 
 const changePasswordTemplate = (token: string) => `some string ${token} ${CLIENT_HOST}`
