+++ services/notifications/src/components/accounts/templates/VerifyEmailTemplate.ts
@@ -1,4 +1,4 @@
-import { CLIENT_HOST } from '../../../configs/AppConfig'
+import { CLIENT_HOST } from '../../../configs/app-config'
 
 const verifyEmailTemplate = (token: string) => `Hi please verify your email ${CLIENT_HOST}/v1/account/verify?token=${token}`
