+++ services/notifications/src/components/index.ts
@@ -1,4 +1,4 @@
-export * as AccountWorker from './accounts/workers/AccountWorker'
-export * as OrganizationWorker from './organization/workers/OrganizationWorker'
-export * as ProfileWorker from './profile/workers/ProfileWorker'
-export * as RoleWorker from './role/workers/RoleWorker'
+export * as AccountWorker from './accounts/workers/account-worker'
+export * as OrganizationWorker from './organization/workers/organization-worker'
+export * as ProfileWorker from './profile/workers/profile-worker'
+export * as RoleWorker from './role/workers/role-worker'
