+++ services/notifications/src/index.ts
@@ -4,8 +4,8 @@ import cluster from 'node:cluster'
 import type { ConsumerConfig } from '@apart-re/notify'
 import { Consumer } from '@apart-re/notify'
 import _ from 'lodash'
-import { SERVER_PORT } from './configs/AppConfig'
-import { handlersByName, WorkersConfig } from './configs/WorkersConfig'
+import { SERVER_PORT } from './configs/app-config'
+import { handlersByName, WorkersConfig } from './configs/workers-config'
 import type { LogLevels } from './utils'
 import { getLogger, TerminationHandler } from './utils'
